Update style.css
added css styling to the buttons after being interactable  through javascript. Will play with the color/theme after finishing to personalize. Noticed major consistency among font, color, padding, margins; May not have to be varied in final projects, but good to mess with to get a feel for what is easier to see or how one may affect another. Somehow found trouble committing this change, had to be done manually.